refactor :  - remove unnecessary interface DataFinder to reduce complexity.  - move DataSetLoader interface and its implementation JsonDataSetLoader to repositories to reflect their role as data source.
