--- src/main/java/com/safetynet/safetynetalertsapi/services/DataFinder.java
@@ -1,7 +1,0 @@
-package com.safetynet.safetynetalertsapi.services;
-
-import com.safetynet.safetynetalertsapi.model.DataSet;
-
-public interface DataFinder {
-	public DataSet processQuery();
-}
